-- Change MSAA 1x MSAA  option
Modified "-- Change MSAA" section to allow for a 1x MSAA option.
This was done because setting the MSAA option to "OFF" just uses the game's default MSAA setting. 
Usage case: Setting to 1x on a game with 4x MSAA  would completely disable the MSAA effect to allow for better framerates.